# MDEV-36184 - To optimise dot_product in Power9 and Power10 architecture