@mikejuliet13 2025-02-24
> > The original dot_product function does undergo auto-vectorisation when compiled with -O3. However, performance analysis has shown that the newly optimised implementation performs better on Power10 and achieves comparable performance on Power9 machines.
> 
> Do you have any numbers? How does your implementation compare with auto-vectorization? Did you benchmark it? On Power 10 and Power 9? Auto-vectorization at -O3 by what compiler version?

I conducted benchmark tests on both Power9 and Power10 machines, comparing the time taken by the original (auto-vectorized) code and the new vectorised code. I used GCC 11.5.0 on RHEL 9.5 operating system with -O3. The benchmarks were performed using a sample test code with a vector size of 4096 and 10⁷ loop iterations.
Here are the average execution times (in seconds) over multiple runs:
Power9:
- Before change: ~16.364 s
- After change: ~16.180 s
- Performance gain is modest but measurable.

Power10:

- Before change: ~8.989 s
- After change: ~6.446 s
- Significant improvement, roughly 28–30% faster.

The final results of the dot product remained the same before and after the change, confirming functional correctness.