@mikejuliet13 2025-04-10
> @mikejuliet13 we should form a single commit, without merge commits and those WiP changes. And also write proper commit message. I can do it if you like, or do you want to do it yourself? git commit message requirements are listed here: https://github.com/MariaDB/server/blob/main/CODING_STANDARDS.md

Hey @svoj,
Could you please squash everything into a single commit