@vuvova 2025-04-14
I've cherry-picked it into my bb-11.8-serg branch, will push it together with other vector-related fixes